+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Data Engineers Snowflake DataOps Utils Project Changelog
 This file contains the changelog for the Data Engineers Snowflake DataOps Utils project, detailing updates, fixes, and enhancements made to the project over time.
 
+## v0.3.9 2024-06-10 - Grant Object For Procedures
+
+* added macro `grant_procedure_usage` to enable the ability to grant usage of a stored procedure to a role
+
 ## v0.3.8.5 2025-08-25 - Grant Usage to Application
 
 * added grant usage to application for `sp_sync_` in the `grant_privileges` macro